pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "openai>=2.8.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "plotly>=6.4.0",
    "pydicom>=3.0.1",
    "scikit-image>=0.25.2",
    "streamlit>=1.51.0",
]
